// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-billing-revenue-currency-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminBillingRevenueCurrencyHttpResponse {
    chargebacks?: number;
    currency?: string;
    earningsTrackedPayments?: number;
    gross?: number;
    net?: number;
    paddleEarnings?: number;
    paddleFees?: number;
    refunds?: number;
    reversals?: number;
    successfulPayments?: number;
    tax?: number;
}