// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-dashboard-metrics-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminDashboardTrendHttpResponse } from './admin-dashboard-trend-http-response';

export interface AdminDashboardMetricsHttpResponse {
    aiTokens?: number;
    payingUsers?: number;
    registrations?: number;
    trend?: Array<AdminDashboardTrendHttpResponse>;
}