// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-dashboard-overview-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminDashboardPeriodHttpResponse } from './admin-dashboard-period-http-response';

export interface AdminDashboardOverviewHttpResponse {
    fromUtc?: string;
    interval?: string;
    pendingReportsNow?: number;
    period?: AdminDashboardPeriodHttpResponse;
    premiumUsersNow?: number;
    previous?: AdminDashboardPeriodHttpResponse | null;
    toUtc?: string;
    totalUsersNow?: number;
}