// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-dashboard-period-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminBillingRevenueCurrencyHttpResponse } from './admin-billing-revenue-currency-http-response';
import { AdminDashboardMetricsHttpResponse } from './admin-dashboard-metrics-http-response';

export interface AdminDashboardPeriodHttpResponse {
    currencies?: Array<AdminBillingRevenueCurrencyHttpResponse>;
    fromUtc?: string;
    metrics?: AdminDashboardMetricsHttpResponse;
    toUtc?: string;
}