// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-dashboard-summary-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminUserHttpResponse } from './admin-user-http-response';

export interface AdminDashboardSummaryHttpResponse {
    activeUsers?: number;
    deletedUsers?: number;
    pendingReportsCount?: number;
    premiumUsers?: number;
    recentUsers?: Array<AdminUserHttpResponse>;
    totalUsers?: number;
}