// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-dashboard-trend-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminDashboardRevenuePointHttpResponse } from './admin-dashboard-revenue-point-http-response';

export interface AdminDashboardTrendHttpResponse {
    aiTokens?: number;
    date?: string;
    registrations?: number;
    revenue?: Array<AdminDashboardRevenuePointHttpResponse>;
}