// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/fasting-telemetry-preset-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface FastingTelemetryPresetHttpResponse {
    checkInRatePercent?: number;
    completedSessions?: number;
    completionRatePercent?: number;
    firstReminderHours?: number | null;
    followUpReminderHours?: number | null;
    presetId?: string;
    savedCheckIns?: number;
    selectionCount?: number;
    startedSessions?: number;
    timingSaveCount?: number;
}