// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/fasting-telemetry-summary-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FastingTelemetryPresetHttpResponse } from './fasting-telemetry-preset-http-response';

export interface FastingTelemetrySummaryHttpResponse {
    averageCompletedDurationHours?: number | null;
    checkInRatePercent?: number;
    completedSessions?: number;
    completionRatePercent?: number;
    generatedAtUtc?: string;
    lastCheckInAtUtc?: string | null;
    lastEventAtUtc?: string | null;
    manualReminderTimingSaves?: number;
    presetReminderTimingSaves?: number;
    reminderPresetSelections?: number;
    reminderTimingSaves?: number;
    savedCheckIns?: number;
    startedSessions?: number;
    topPresets?: Array<FastingTelemetryPresetHttpResponse>;
    windowHours?: number;
}